[refactor][client] Publish combined shaded client under separate coordinates
Keep pulsar-client-v5-all unshaded and publish pulsar-client-v5-shaded with its own dependency-reduced POM. Remove classifier-specific publication logic, validate plain Maven and Gradle dependencies, and document Maven classifier limitations.
