+++ ARCHITECTURE.md
@@ -21,7 +21,7 @@ accordingly:
 - **`pulsar-client-api`, `pulsar-client-admin-api`** — public, backward-compatible interfaces only.
   `pulsar-client-api-v5` / `pulsar-client-v5` are the newer V5 client API (PIP-466/468).
   Applications using either v4 or v5 can use [`pulsar-client-v5-all`](pulsar-client-v5-all/README.md)
-  for the client and admin together: an unshaded aggregate by default, or a shaded `all` classifier.
+  for the unshaded client and admin together, or `pulsar-client-v5-shaded` for the combined shaded jar.
 - **`pulsar-client` (`:pulsar-client-original`)** — the Java client implementation
   (producer/consumer/reader, connection pooling). `pulsar-client-admin` implements the admin REST
   client.
