+++ pulsar-client-api-v5/src/main/java/org/apache/pulsar/client/api/v5/internal/PulsarClientProvider.java
@@ -119,7 +119,8 @@ public interface PulsarClientProvider {
                 .findFirst()
                 .orElseThrow(() -> new IllegalStateException(
                         "No PulsarClientProvider found on the classpath. "
-                                + "Add pulsar-client-v5 or pulsar-client-v5-all to your dependencies."));
+                                + "Add pulsar-client-v5, pulsar-client-v5-all, "
+                                + "or pulsar-client-v5-shaded to your dependencies."));
 
         private Holder() {
         }
