+++ pulsar-client-shaded/build.gradle.kts
@@ -19,7 +19,6 @@
 
 plugins {
     id("pulsar.public-java-library-conventions")
-    id("pulsar.shadow-conventions")
     id("pulsar.client-shade-conventions")
 }
