+++ pulsar-client-v5-shaded/build.gradle.kts
@@ -24,7 +24,7 @@ plugins {
 
 dependencies {
     // Bundled into the shaded jar (kept on the runtime classpath so shadowJar packs them):
-    implementation(project(":pulsar-client-original")) {
+    implementation(project(":pulsar-client-v5")) {
         // fastutil is bundled via :pulsar-client-fastutil-minimized, which packs only the
         // (unrelocated) fastutil classes actually used instead of the full ~25MB jar.
         exclude(group = "it.unimi.dsi", module = "fastutil")
