+++ settings.gradle.kts
@@ -204,6 +204,7 @@ include("pulsar-broker-fastutil-minimized")
 include("pulsar-client-shaded")
 include("pulsar-client-all")
 include("pulsar-client-v5-all")
+include("pulsar-client-v5-shaded")
 include("pulsar-client-admin-shaded")
 
 // Tier 11 — distribution (server is in core-modules)
@@ -269,7 +270,7 @@ project(":tests:integration").projectDir = file("tests/integration")
 include("tests:pulsar-client-shade-test")
 project(":tests:pulsar-client-shade-test").projectDir = file("tests/pulsar-client-shade-test")
 include("tests:pulsar-client-admin-v5-test")
-include("tests:pulsar-client-v5-all-shade-test")
+include("tests:pulsar-client-v5-shade-test")
 include("tests:pulsar-client-v5-all-test")
 include("tests:pulsar-client-admin-shade-test")
 project(":tests:pulsar-client-admin-shade-test").projectDir = file("tests/pulsar-client-admin-shade-test")
