+++ tests/pulsar-client-v5-shade-test/src/test/java/org/apache/pulsar/tests/integration/ClientClasspathTest.java
